code splitting to a RCC form submitting select room
